sanitizacao e senha em hash implementados

+++ redefinir-senha.php
@@ -65,10 +65,10 @@ if (strtotime($user[0]) <= time()) {
             <p>Senhas fortes incluem números, letras e sinais de pontuação.</p>
             <div id="input-group">
                 <label><strong>Digite sua nova senha:</strong></label>
-                <input type="password" maxlength="15" name="senha" placeholder="máx de 15 caracteres" />
+                <input type="password" maxlength="15" name="senha" placeholder="mínimo de 8 caracteres." />
                 <br/><br/>
                 <label><strong>Confirme sua nova senha:</strong></label>
-                <input type="password" maxlength="15" name="confirmaSenha"/>
+                <input type="password" maxlength="15" name="confirmaSenha" placeholder="mínimo de 8 caracteres."/>
             </div>
             <br/><br/>
             <input type="submit" value="REDEFINIR SENHA" class="btn-recuperacao">
@@ -83,18 +83,27 @@ if (strtotime($user[0]) <= time()) {
             $confirmaSenha = $_POST['confirmaSenha'];
 
             if (!empty($senha) && !empty($confirmaSenha)) {
-                if ($senha == $confirmaSenha) {
-
-                    $exec5 = $pdo->prepare("UPDATE tblUsuario SET senha_usuario = :senha, token_rec_senha = null, dt_expiracao_token = null WHERE token_rec_senha = :token_hash;");
-                    $exec5->bindValue(":token_hash", $token_hash);
-                    $exec5->bindValue(":senha", $senha);
-                    $exec5->execute();
-
-                    header("location: index.html");
-                    echo "<script>window.location.replace('index.html');</script>";
-                    die();
+                if (strlen($senha) >= 8 && strlen($confirmaSenha) >= 8) {
+                    if ($senha == $confirmaSenha) {
+                        if (!preg_match('/[\s\'"=*;()-]/', $senha)) {
+                            $senha_hash = password_hash($senha, PASSWORD_DEFAULT);
+    
+                            $exec5 = $pdo->prepare("UPDATE tblUsuario SET senha_usuario = :senha, token_rec_senha = null, dt_expiracao_token = null WHERE token_rec_senha = :token_hash;");
+                            $exec5->bindValue(":token_hash", $token_hash);
+                            $exec5->bindValue(":senha", $senha_hash);
+                            $exec5->execute();
+        
+                            header("location: index.html");
+                            echo "<script>window.location.replace('index.html');</script>";
+                            die();
+                        } else {
+                            echo "<script>alert('A nova senha contém caracteres inválidos.');</script>";
+                        }
+                    } else {
+                        echo "<script>alert('As senhas não coincidem');</script>";
+                    }
                 } else {
-                    echo "<script>alert('As senhas não coincidem');</script>";
+                    echo "<script>alert('A nova senha deve conter ao menos 8 caracteres.');</script>";
                 }
             } else {
                 echo "<script>alert('Preencha todos os campos');</script>";
